requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
catboost==1.2.7
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
supabase==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
huggingface_hub==0.27.0
hf_xet==0.1.0
understatapi